Move out searchContext into its own module to avoid a hard require to the code intel APIs
